Refatored  extract zipDirectory into dedicated functions
