+++ src/commands/sfpowerkit/org/trigger/deactivate.ts
@@ -10,11 +10,12 @@ import xml2js = require('xml2js');
 import util = require('util');
 // tslint:disable-next-line:ordered-imports
 var jsforce = require('jsforce');
-var path = require('path')
-var unzipper = require('unzipper')
-var archiver = require('archiver');
+var path = require('path');
+
 import { checkRetrievalStatus } from '../../../../shared/checkRetrievalStatus';
 import { checkDeploymentStatus } from '../../../../shared/checkDeploymentStatus';
+import { extract } from '../../../../shared/extract';
+import { zipDirectory } from '../../../../shared/zipDirectory';
 
 
 
@@ -155,32 +156,6 @@ export default class Deactivate extends SfdxCommand {
 
 }
 
-const extract = (location: string) => {
-  return new Promise((resolve, reject) => {
-    fs.createReadStream(`./${location}/unpackaged.zip`)
-      .pipe(unzipper.Extract({ path: `${location}` }))
-      .on('close', () => {
-        resolve();
-      })
-      .on('error', error => reject(error));
-  });
-};
-
-const zipDirectory = (source, out) => {
-  const archive = archiver('zip', { zlib: { level: 9 } });
-  const stream = fs.createWriteStream(out);
-
-  return new Promise((resolve, reject) => {
-    archive
-      .directory(source, false)
-      .on('error', err => reject(err))
-      .pipe(stream)
-      ;
-
-    stream.on('close', () => resolve());
-    archive.finalize();
-  });
-};
